+++ src/main/java/com/Jobsteer/Jobsteer/repositories/ResumeRepository.java
@@ -3,5 +3,5 @@ package com.Jobsteer.Jobsteer.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.Jobsteer.Jobsteer.entities.Resume;
 
-public interface ResumeRepository extends JpaRepository<Resume, Long> {
+public interface ResumeRepository extends JpaRepository<Resume, Integer> {
 }
